mysite/wallet/serializers.py
from rest_framework import serializers
from .models import Task

# class userProfileSerializer(serializers.ModelSerializer):
#     user = serializers.StringRelatedField(read_only=True)

#     class Meta:
#         model=userProfile
#         fields='__all__'

class TaskSerializer(serializers.ModelSerializer):
	class Meta:
		model = Task
		fields = '__all__'