requirements.txt
Django==2.1.4
djangorestframework==3.7.7
django-filter==2.0.0
